Move StoreByBucket tests to separate test file.
