+++ searchlib/src/tests/docstore/store_by_bucket/FILES
@@ -1,0 +1,1 @@
+store_by_bucket_test.cpp
